Moved sorting to comments model
