+++ server/shared/mail/index.js
@@ -2,11 +2,9 @@
 
 const { renderHtml, renderText } = require('./render');
 const { origin } = require('../../../config/server');
-const { getOutlineLevels } = require('../../../config/shared/activities');
 const email = require('emailjs');
 const logger = require('../logger');
 const path = require('path');
-const find = require('lodash/find');
 const pick = require('lodash/pick');
 const Promise = require('bluebird');
 
@@ -72,9 +70,9 @@ function getConfig(server) {
   ]);
 }
 
-function commentsList({ user, comments, schema, days }) {
+function commentsList({ user, comments, days }) {
   const recipient = user;
-  const data = { comments: sortComments({ comments, schema }), recipient, days };
+  const data = { comments, user, days };
   const html = renderHtml(path.join(templatesDir, 'comments.mjml'), data);
   const text = renderText(path.join(templatesDir, 'comments.txt'), data);
   logger.debug({ recipient, sender: EMAIL_ADDRESS }, '📧  Sending comments email to:', recipient);
@@ -85,24 +83,6 @@ function commentsList({ user, comments, schema, days }) {
     text,
     attachment: [{ data: html, alternative: true }]
   });
-}
-
-function sortComments({ comments, schema }) {
-  const structure = getOutlineLevels(schema);
-  const sortedComments = [];
-  let temp = [];
-  comments.forEach((comment, i) => {
-    const { label, color } = find(structure, { type: comment.activity.type });
-    comment.activity.label = label;
-    comment.activity.color = color;
-    if (i && comment.activityId !== comments[i - 1].activityId) {
-      sortedComments.push(temp);
-      temp = [];
-    }
-    temp.push(comment);
-  });
-  sortedComments.push(temp);
-  return sortedComments;
 }
 
 module.exports = {
